requirements.txt
streamlit
streamlit-chat
requests
Pillow
matplotlib
google-generativeai
python-dotenv
selenium
beautifulsoup4
